commit(1.2): Пробую создать массив
