+++ src/mock/offers.js
@@ -1,0 +1,164 @@
+import {getRandomInteger} from './mockData.js';
+
+export const offersData = [
+  {
+    type: 'flight',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'bus',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'train',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'ship',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'drive',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'check-in',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'sightseeing',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+  {
+    type: 'restaurant',
+    offers: [
+      {
+        id: 0,
+        description: 'economy class',
+        price: getRandomInteger(20, 150),
+      },
+      {
+        id: 1,
+        description: 'bissnes class',
+        price: getRandomInteger(100, 400),
+      },
+      {
+        id: 2,
+        description: 'extra service',
+        price: getRandomInteger(10, 100),
+      }
+    ]
+  },
+];
